# Speedup compile time for rustpython-parser