--- a/parser/regen_lalrpop.sh
+++ b/parser/regen_lalrpop.sh
@@ -1,0 +1,1 @@
+lalrpop src/python.lalrpop