Saves about half a minute on a clean build, see timing reports:

[Before](https://gistcdn.githack.com/coolreader18/65e901aedc12eca06c7df41df6910373/raw/7ebbb40938bd4d91a7e04aa87298e0a123c1d0dd/cargo-timing-20220422T193333Z.html)



Targets: | rustpython 0.1.2 (lib, bin "rustpython")
-- | --
Profile: | release
Fresh units: | 0
Dirty units: | 345
Total units: | 345
Max concurrency: | 8 (jobs=8 ncpu=8)
Build start: | 2022-04-22T19:33:33Z
Total time: | 177.4s (2m 57.4s)

[After](https://gistcdn.githack.com/coolreader18/65e901aedc12eca06c7df41df6910373/raw/7ebbb40938bd4d91a7e04aa87298e0a123c1d0dd/cargo-timing-20220422T220937Z.html)



Targets: | rustpython 0.1.2 (lib, bin "rustpython")
-- | --
Profile: | release
Fresh units: | 0
Dirty units: | 320
Total units: | 320
Max concurrency: | 8 (jobs=8 ncpu=8)
Build start: | 2022-04-22T22:09:37Z
Total time: | 153.6s (2m 33.6s)