@coolreader18 2022-04-22
Achh, newline issues on windows.