@fanninpm 2022-04-23
# TAKE MY BENCHMARKS WITH A GRAIN OF SALT

HP Elitebook 745 G3
Manjaro Stable (2022-04-15 update)
Rust v1.60.0
\*processes are running in the background\*
~~listening to some [Bach](https://youtu.be/qBOJDTvG0DQ) on another computer~~

## Current `main` (c103fe13)

| Targets: | rustpython 0.1.2 (lib, bin "rustpython") |
|---|---|
| Profile: | release |
| Fresh units: | 0 |
| Dirty units: | 375 |
| Total units: | 375 |
| Max concurrency: | 4 (jobs=4 ncpu=4) |
| Build start: | 2022-04-23T01:04:52Z |
| Total time: | 726.9s (12m 6.9s) |
| rustc: | rustc 1.60.0 (7737e0b5c 2022-04-04) Host: x86_64-unknown-linux-gnu Target: x86_64-unkown-linux-gnu |
| Max (global) rustc threads concurrency: | 0 |

## This branch (2918395)

| Targets: | rustpython 0.1.2 (lib, bin "rustpython") |
|---|---|
| Profile: | release |
| Fresh units: | 0 |
| Dirty units: | 357 |
| Total units: | 357 |
| Max concurrency: | 4 (jobs=4 ncpu=4) |
| Build start: | 2022-04-23T01:38:57Z |
| Total time: | 575.9s (9m 35.9s) |
| rustc: | rustc 1.60.0 (7737e0b5c 2022-04-04) Host: x86_64-unknown-linux-gnu Target: x86_64-unkown-linux-gnu |
| Max (global) rustc threads concurrency: | 0 |